moving PHIR capabilities to pecos-phir crate
