+++ crates/pecos-engines/src/lib.rs
@@ -12,11 +12,7 @@ pub use engines::{
     hybrid::HybridEngine,
     monte_carlo::MonteCarloEngine,
     noise::{DepolarizingNoiseModel, NoiseModel, PassThroughNoiseModel},
-    phir::PHIREngine,
     quantum::QuantumEngine,
     quantum_system::QuantumSystem,
 };
 pub use pecos_core::errors::PecosError;
-
-// Re-export engine setup functions
-pub use engines::classical::setup_phir_engine;
