+++ crates/pecos-phir/README.md
@@ -1,0 +1,160 @@
+# PECOS High-level Intermediate Representation (PHIR)
+
+This crate provides parsing and execution capabilities for the PECOS High-level Intermediate Representation (PHIR), a
+JSON-based format for representing quantum programs in the PECOS quantum simulator framework.
+
+## Overview
+
+PHIR is designed to:
+
+- Provide a human-readable representation of quantum circuits
+- Support a mix of quantum and classical operations
+- Allow for deterministic execution of quantum programs
+- Serve as an intermediate layer between high-level languages and lower-level simulators
+
+## Usage
+
+### Basic Example
+
+```rust
+use pecos_phir::PHIREngine;
+use pecos_engines::core::shot_results::OutputFormat;
+use std::path::Path;
+
+// Load a PHIR program from a file (v0.1 implementation)
+let engine = PHIREngine::new(Path::new("examples/bell.json"))?;
+
+// Process the program
+let results = engine.process(())?;
+
+// Format the results
+let formatted_results = engine.get_formatted_results(OutputFormat::PrettyJson)?;
+println!("{}", formatted_results);
+```
+
+### Using with Automatic Version Detection
+
+```rust
+use pecos_phir::setup_phir_engine;
+use pecos_engines::{MonteCarloEngine, engines::noise::DepolarizingNoiseModel};
+use std::path::Path;
+
+// Create a classical engine from a PHIR program file
+// The version will be automatically detected from the file
+let classical_engine = setup_phir_engine(Path::new("examples/bell.json"))?;
+
+// Run the program with a noise model
+let noise_model = Box::new(DepolarizingNoiseModel::new_uniform(0.01));
+let results = MonteCarloEngine::run_with_noise_model(
+    classical_engine,
+    noise_model,
+    100, // shots
+    2,   // workers
+    None // seed
+)?;
+
+println!("{}", results);
+```
+
+### Explicit Version Selection
+
+```rust
+// For specific version implementations
+use pecos_phir::setup_phir_v0_1_engine;
+use std::path::Path;
+
+// Explicitly use v0.1 implementation
+let engine = setup_phir_v0_1_engine(Path::new("examples/bell.json"))?;
+```
+
+## PHIR File Format
+
+PHIR files are JSON documents with the following structure:
+
+```json
+{
+  "format": "PHIR/JSON",
+  "version": "0.1.0",
+  "metadata": {
+    "description": "Example PHIR program"
+  },
+  "ops": [
+    {
+      "data": "qvar_define",
+      "data_type": "qubits",
+      "variable": "q",
+      "size": 2
+    },
+    {
+      "data": "cvar_define",
+      "data_type": "i64",
+      "variable": "m",
+      "size": 2
+    },
+    {"qop": "H", "args": [["q", 0]]},
+    {"qop": "CX", "args": [["q", 0], ["q", 1]]},
+    {"qop": "Measure", "args": [["q", 0]], "returns": [["m", 0]]},
+    {"qop": "Measure", "args": [["q", 1]], "returns": [["m", 1]]},
+    {"cop": "Result", "args": ["m"], "returns": ["c"]}
+  ]
+}
+```
+
+See the [specification](specification/v0.1/spec.md) for more details.
+
+## Validation and Execution
+
+This crate provides:
+
+1. **Validation**: Rust-based parsing and validation of PHIR programs against the specification
+2. **Execution**: Full integration with PECOS for running PHIR programs on quantum simulators
+3. **Error Handling**: Detailed error messages for both validation and runtime errors
+
+For alternative validation, the [Python Pydantic PHIR validator](https://github.com/CQCL/phir) is also available.
+
+> **Note**: Work is currently in progress to extend the PHIREngine to support the full PHIR specification. Some
+> advanced features may not be fully implemented yet. The specification itself is also evolving - the "Result"
+> command for exporting measurement results is being added as part of a v0.1.1 specification update.
+
+## Supported Operations
+
+### Quantum Operations
+
+- Single-qubit gates: `H`, `X`, `Y`, `Z`
+- Rotations: `RZ`, `R1XY`
+- Two-qubit gates: `CX` (CNOT), `SZZ` (ZZ interaction)
+- Measurement: `Measure`
+
+### Classical Operations
+
+- `Result`: Used to export measurement results
+
+## Versioning
+
+This crate implements a versioning strategy to handle multiple versions of the PHIR specification. See
+[VERSIONING.md](VERSIONING.md) for details on how versions are managed.
+
+### Available Versions
+
+- **v0.1**: The initial version, supporting basic quantum operations, variable definitions, and classical exports.
+  - Specification: [specification/v0.1/spec.md](specification/v0.1/spec.md)
+  - Feature flag: `v0_1` (enabled by default)
+
+### Feature Flags
+
+You can control which PHIR versions are included in your build using Cargo feature flags:
+
+```toml
+# Default: only include v0.1
+pecos-phir = { version = "0.1" }
+
+# Explicitly select a specific version
+pecos-phir = { version = "0.1", default-features = false, features = ["v0_1"] }
+
+# Include all available versions
+pecos-phir = { version = "0.1", features = ["all-versions"] }
+```
+
+## License
+
+This crate is licensed under the Apache License, Version 2.0.
