+++ crates/pecos/src/program.rs
@@ -1,6 +1,7 @@
 use log::debug;
 use pecos_core::errors::PecosError;
 use pecos_engines::ClassicalEngine;
+use pecos_phir::setup_phir_engine;
 use std::path::{Path, PathBuf};
 
 /// Represents the types of programs that PECOS can execute
@@ -150,7 +151,7 @@ pub fn setup_engine_for_program(
 
     match program_type {
         ProgramType::QIR => crate::engines::setup_qir_engine(program_path, None),
-        ProgramType::PHIR => pecos_engines::setup_phir_engine(program_path),
+        ProgramType::PHIR => setup_phir_engine(program_path),
         ProgramType::QASM => crate::engines::setup_qasm_engine(program_path, seed),
     }
 }
